qt: networkmediamodel: execute media_tree callback in Qt thread
Delegate media_tree callback handling to the Qt thread, so that
NetworkMediaModel's internal is locked and race-free.

It needs Qt::QueuedConnection because those callbacks are either called
from the service discovery thread (normal runtime behaviour) or from the
Qt thread itself when it adds the listener (initialization). Without
Qt::QueuedConnection, the lambda won't be queued in the later case as it
is run in the same thread, and will be executed in place, which in turn
leads to a double lock.

Signed-off-by: Thomas Guillem <thomas@gllm.fr>