opengl: fix matrix multiplication order
The orientation matrix must be applied on the input coordinates, before
the (internal) Android transformation, which depends on how the texture
is actually stored.

For example, if the orientation matrix represents a vertical flip,
"vertical" is only meaningful on the input coordinates (if the Android
transformation rotates by 90°, the vertical flip would result in an
horizontal flip instead).

The first version was actually correct:
<https://mailman.videolan.org/pipermail/vlc-devel/2019-December/130153.html>

Signed-off-by: Thomas Guillem <thomas@gllm.fr>