contrib: detect the pkg-config variant to use when cross-compiling
And use it for all contribs detection and building.

on Debian, the x86_64-w64-mingw32-pkg-config may exist and run but give an
error when actually trying to find a package because the package architecture
doesn't exist for that target. So we have to test it thoroughly.

If we revert to pkg-config when cross-compiling we also force the
PKG_CONFIG_LIBDIR, otherwise we leave it untouched as the system/environment may
provide more that we know.

Both /usr/$(HOST)/lib/pkgconfig and /usr/lib/$(HOST)/pkgconfig variants exist,
at least on Debian.