Update vlc.desktop.in
There is no word "redor" in Romanian language. It supposes to be the noun for the Romanian verb "a reda" (to play). However players are called as in English in Romanian. The only difference would be the order of the words. It sounds Romanian to say "un player media".

There can be a discussion to use the word "emițător" instead of "player" but the verb form of the noun is more correctly translated to "to stream".

Please see also:
https://dexonline.ro/definitie/redor
https://dexonline.ro/definitie/player