udp: do not set SO_REUSEPORT on Linux (refs #23201)
VLC set the SO_REUSEPORT socket option on BSD so that multiple UDP
sockets can receive multicast packets on the same port. On Linux,
that behaviour is covered by SO_REUSEADDR.

In "recent" times, Linux added a SO_REUSEPORT option that does
something else - enable load-balancing, which is not the intended
behaviour.