direct3d11: move the external device handling in the common code
The D3D11 display module may be used without a decoder device (software decoding)
and we still need to use the device from the host app.

The allocated d3d11_decoder_device_t structure is freed with the object that
created it. But D3D11_ReleaseDevice() still need to be called.