package/win32: build.sh: don't run bootstrap if it was run once
If the configure.ac changes, autotools will notice and rerun the necessary
commands to build properly. Same thing if a Makefile.am is modified.

If one of the local .m4 file is modified, it won't pick the difference. The
user will need to run the bootstrap manually (or remove the configure file).