contrib: always use $(PKG_CONFIG) to find available dependencies
When cross-compiling the PKG_CONFIG is either forced by the environment, to find
what the user wants us to find, or we detect a usable one and set common system
pathes where we might find usable packages.

It's possible to disable the use of PKG_CONFIG like it used to be by setting
PKG_CONFIG=false in the environment, when building contribs.

We make sure PKG_CONFIG_LIBDIR is used when we set its value.