input: fix seeking for external subtitles when changing chapters
This commit allows external subtitles to be shown correctly when changing to
previous chapters. Currently external subtitles would not be shown until the
played chapter was reached again. Seeking to different time points does not
have this issue. This seems to be the case because there is no SlaveSeek call
for INPUT_CONTROL_SET_SEEKPOINT_PREV, which this patch now adds.

Signed-off-by: Thomas Guillem <thomas@gllm.fr>