codec: omxil: use switch lut for string conversion
Switch are optimized at least to the same form anyway, and they provide
better warnings in case of missing items as well as preventing typos.

It also fixes a cast warning as OMX_STATETYPE was used instead of the
enum ERRORTYPE in the error to string function, and adds missing items
from the enums.

Signed-off-by: Thomas Guillem <thomas@gllm.fr>