vout/caopengllayer: fix scaling when changing displays
Changing from a retina screen to a non-retina screen changes the
scale of the layer, so from point of view of the OpenGL code, the
size of the surface changes. Therefore we need to signal this change
so that the size is adapted correctly.

This is not a perfect solution as the size event will arrive too late
so for a few frames the wrong size can still be seen before it switches
to the new size. Ideally this would be handled like the live resize
but so far I have not found a good solution for how to do that.

Forward-port from 3.0.x branch.
Cherry-picked from commit 57401f907ca98bfd37e73577014642fd16c123bb.

Signed-off-by: Alexandre Janniaux <ajanni@videolabs.io>