configure: Check for if_nametoindex in iphlpapi.h
If the toolchain defaults to a newer version of windows, this
function is available there, but the previous configure checks
didn't notice it. Force testing with both compiling and linking,
as the function exists in import libraries, but only is visible
in headers since Vista.

On the master branch, the function was taken into use
unconditionally in c0bf5aeefc611f8d1e1ace6111c88956c0a8fa6c.