contrib: wine-headers: install dxva.h headers from wine
They are already in mingw64 and might not be matching (dxva vs dxva2api). We
should only use one or the other but not a file from one with the other file
from the other.

New headers will also be needed soon to handle HEVC 444 or AV1 in DXVA and
since we already install the headers from here we could patch them from here.

Also dxva2api.h is not just for D3D9 and doesn't need special handling.