libvlc: comments and doxygen cleanups
Un-doxygens a lot of doc comments that were already present in the
header file and removes the redundant long descriptions as those are
just prone to get outdated and never updated along with the proper
doc comments in the header.

Change the doc comments for the internal functions to proper
doxygen with the functions marked as internal.