CSSGrammar: refactor lex/parse-param
Scanner is used by both lexer and parser, so %param allows binding to
both, and css_parser is only used by the parser so there is no need to
declare it in the yylex function. As the function was not correctly
exposed to the lexer code, it was leading to different prototype for
both generated compile unit.

(cherry picked from commit b8a4ccee33423f4b1a6d48ca2def1a06a6ae93d6)