opengl: simplify renderer format
The interop may modify its own copy of the video_format_t. These changes
must be reported to the core (by writing to the provided
video_format_t).

But as a hack, the Android interop currently flips the orientation to
avoid a vertical flip in the renderer, since it provides its own
transform matrix. This change must not be reported to the core (the core
must not flip the input), so the orientation change is reverted. A
further refactor will remove this hack later.

Meanwhile, handle the format to report to the core outside of the
renderer, so that the renderer is not involved in this hack. This paves
the way to pass only the interop to the renderer, without an additional
video_format_t instance.

Signed-off-by: Alexandre Janniaux <ajanni@videolabs.io>