placebo: update for new tone mapping desaturation algo
This was introduced in API version 10, and refactors the tone mapping
desaturation curve into a new, more tunable algorithm that has some
different behavior. In particular, it allows us to simulate exactly the
"hollywood" style of tone mapping, so we document those special values
specifically.

This is a manual backport of b579384a239683232f5558216cc03c6ae5a5dc38