qt: player_controller: check item before usage
Request like restoration of the last playback position are emitted from
the UI, and the player event signalling that the input item changed is
emitted from a player thread and asynchronously processed later by the
UI thread. It means that the UI could request the player to do
something which would be executed on the wrong item, or no item at all.

Store the input item currently acknowledged by the UI thread and check
it against the current item being played by the player when the player
requests are emitted by the UI. If they don't match, avoid processing
the requests.

Some requests are not concerned by these change, either because they are
able to affect the following media (change rate) or because they could
potentially be triggered near the end of a media and already handle a
nullptr input (record, a-b loop).

Fixes #24347