opensles_android: fix integer print format
SLresult is defined as an SLuint32 and might not be represented as an
unsigned long int. It is raising warnings when compiling for android
aarch64, so cast it to ensure it's an uint32_t and display it with
PRIu32 instead.

block_t->i_buffer is size_t too and must be displayed with %zu.