Remove --daemon
Forking the process after LibVLC has already used threading is not
possible, so this has not been working for years, if it ever worked.

Some users do run VLC in the background, but they generally have their
own wrappers, and/or use dedicated job management such as upstart or
systemd. daemon() has the fatal problem that it prevents checking
for error returns (early exit) as well as the standard error.

This also removes --pidfile since it's a no-op without --daemon.