soundcloud.lua: dynamically extract API magic
It seems that the client_id API magic now gets invalidated much faster,
rendering the current model, where it is hard-coded into the script,
not viable anymore. Instead, this fetches and parses javascript assets
linked from the web page to search and extract the up-to-date magic from
them.

This is significantly slower as for now it fetches all of several
indiscriminate javascript assets, but at least it works, and should not
require that kind of maintenance anymore.

(cherry picked from commit 6abb32a9ba076de8662a53889926280e6d80f5e1)
Signed-off-by: Pierre Ynard <linkfanel@yahoo.fr>