vout: win32: Remove an unused local property key definition
This one used to be missing from mingw-w64 headers, but was added there
yesterday in git master.

If this one actually is used, and we can't be sure of whether it exists
in system headers, we could define it ourselves, but with a unique
prefix. But in this case it simply seems to be unused.

This fixes compilation with the very latest mingw-w64 headers.

(cherry picked from commit 5398108a0c52d20ba0f657339a4465136b80e4e5)