configure: Check the toolchain default _WIN32_WINNT in addition to a command line override
Since 255e2ce27, we try not to override _WIN32_WINNT in case it already
is defined on the command line to a higher value. However, if it isn't
specified on the command line, but the toolchain headers default to
a newer version, we should also honor it and keep that version instead
of forcing a lower version here. (If the toolchain defaults to a newer
version, runtime libs of the toolchain may rely on such a new version
anyway, so forcing a lower target within VLC might be useless.)

EDIT: Adapted to a different default version in the 3.0 branch.

(cherry picked from commit 0d65fe038a76d3686d2e37c5b522fb24b92d8157)