sap: remove SAPv0 support
This removes non-strict mode, in other words support for non-standard
version 0 of the protocol. This means releases of the session directory
tool (sdr) from before October 1996.

SAP was only standardized as version 2 in 2000. SAP support was added
to VLC in 2003; VLC and miniSAPserver have always sent packets following
the standard (except for the first 6 days of development where a
completely custom packet format was used, never seen in a release).

We retain backward compatibility with version 1 (sdr releases from 1996
to 2000) because there are no relevant differences with version 2.
Version 0 packets however lack the hash value, which *is* troublesome.