contrib: lua: autodetect lua >= 5.1
When building on ArchLinux currently, the contrib was built even if
lua5.3 is installed. As at least Archlinux and Debian are exposing both
a luaX.Y.pc and lua.pc file for the latest supported version of lua,
detecting that the system version is correct through lua.pc should
also ensure that the latest compatible version is installed and used
instead.

The other checks are still needed as you can for example install lua5.2
(which will expose a lua5.2.pc file but no lua.pc file) without having
to install lua (5.3).

It is worth noticing that being able to use lua >= 5.1 doesn't mean any
version of lua >= 5.1 would be compatible. The source code is completely
compliant to those version and native builds will have no issues with
using a more recent version, but cross-compilation environment for 32bit
environment from a 64bit one won't and need the appropriate version to
be installed.