wayland: xdg-shell: use vlc_cond_t instead of dispatch
When using wl_display_dispatch, there is a race condition with the main
loop, in which the condition `(!sys->wm.configured)` would have been
checked as true, but the event dispatched from the main thread after
that so wl_display_dispatch would block indefinitely.

Said differently, wl_display_dispatch should not be called from a
different thread than the main thread (in wayland terms).

By using a condition variable instead, we let the main thread unblock
the vout_window_Enable and prevent this deadlock when the callback for
the configured event is called. In addition, it adds the correct lock
around the updates and reads of the wm.configured boolean.

Fixes #24586