oldmovie: fix integer usage
Use size_t when indexing and long long for values depending on time.
Warnings on abs with long int are fixed and comparison between signed
and unsigned integer are avoided at the same time.

It also has the benefit of easing the reading a bit by using shorter
names or splitting some operations.