qt: use local-path for ui header wrapper
Those files could generated in implementation-specified location by
build systems, as it's the behaviour of qmake based build system.

In particular, meson implementation provides the same behaviour and
it will be handled automatically, thus having a full path makes the
file not found with meson.

Use local-path and specify all include directories to simulate this.
New files must not have the same filename as existing ones.