queue: add generic queue/FIFO helpers
Currently, the FIFO helpers are tied to block_t. This has proven problematic
in the past: some code uses block_t only for the sake of block_fifo_t. It
will get worse with the introduction of separate vlc_frame_t and vlc_data_t.

This provides a generic C implementation of a thread-safe queue. It is very
much meant to be wrapped with more type-specific helpers.