rtp: remove support for RTP framing over TCP
This is specified in RFC 4571. However there are no standard ways to
negotiate it without COMEDIA (RFC 4572). Since COMEDIA implies
SDP Offer/Answer, it cannot be used with VLC so far.

Given that there is no and never was a way to send such stream with the
VLC streaming output, it is pretty safe to assume that this was never
used. To make matters worse, the receival code never worked on Windows.

Remove this at least for the time being.

Note: This was meant for transfering RTP/RTCP over dedicated TCP
streams. This was *not* meant to do "interleaved" TCP whereby RTP is
sent over the same TCP stream as RTSP controls.