configure: check if if_nametoindex is available on Windows
It may link with iphlpapi but be forbidden in UWP.

This test relies on the fact that -Werror-implicit-function-declaration works
in the compiler. And that's why it's done after it's enabled.

It's only used for the --miface option to set the preferred multicast interface.

(cherry picked from commit 424e8991d357e948627bd8ce8872ce7e8a5de6a7) (edited)

edited:
- if_nametoindex was already detected, so just moved the code after the
 -Werror-implicit-function-declaration test

Signed-off-by: Steve Lhomme <robux4@ycbcr.xyz>