contrib: vpx: remove strip step and fix debug symbols
Without CONFIG_DEBUG set, libvpx is calling `$(STRIP) --strip-debug`
on the final library and removes the debug symbols from it. With the
variable set, it becomes a simple copy operation.

In addition, debug symbols should be enabled for all builds since we
strip in the end, so as to have the debug symbols in a separate file
when doing releases. In case we build with optimization, remove the
assertions.

(cherry picked from commit 52e22b96239cc3784aa13495be92f5f3ce591bf1) (rebased)
rebased:
- the code before is slightly different