dvdnav: Improve user information about insufficient permissions
The mentioned security setting is relevant for accessing
external media like RAW block devices (/dev/xxx), which is one of
the main use cases for this module (modern Macs do not have internal
optical drives anymore).

Probe for this case explicitly to inform the user how to get
access.

This is only relevant starting with macOS Catalina.