vout: win32: drawable: use an internal HWND to detect size changes
In the past the display module was checking if the size of the host HWND
changed via the Manage callback. Now this callback doesn't exist anymore and
the window module is responsible for reporting its size.

In the case of embedded-window we need to read the size of the parent as we
don't get any events telling us it's resized, nor that we should be resized
(since apps using set_hwnd have no way to know the HWND we use internally to
resize it).