contrib: gnutls: fix forbidden calls in Winstore builds
Backport a UNICODE fix that was merged after the GnuTLS code freeze.

Forbid call to GetHandleInformation(). fcntl never goes in that code section in
Winstore builds. It's only used for linux random number generation.

Do not allow calling SecureZeroMemory which is forbidden which is forbidden
and not available as an inline function in some cases.