coreaudio: fix invalid delay after a unpause
The render host time was not updated while paused.

This caused the first time_get(), after a unpause, to return a delay way too
early (corresponding to the pause time). This could happen only when the ca
render callback was not triggered between an unpause and a time_get.

This invalid delay caused the insertion of a long silence, that was not
interruptible, hence the impression of a deadlock.

Fixes #24668

(cherry picked from commit f9fce13591bf3f99c4d9401fe3ba67fb30b0fb75)
Signed-off-by: Thomas Guillem <thomas@gllm.fr>