win32: use windowsappcompat instead of winstorecompat
This is the proper counterpart to windowsapp.

Now that we have a proper Docker image to build it:
registry.videolan.org/vlc-debian-llvm-uwp:20200603145315

A recent mingw64 8 (unreleased) is needed to make use of this. It's available
in our Docker images and in msys2 (although it's using msvcrt so it will
probably fail to link properly)

The forced -lwindowsappcompat is added like the other LDFLAGS in configure.ac.

(cherry picked from commit 6e8effb678a6a65325235a7336636aed517d4333) (edited)

edited:
- this branch doesn't set winstorecompat in configure.ac but relies on the toolchain

Signed-off-by: Steve Lhomme <robux4@ycbcr.xyz>