package/win32: build.sh: do not link with ucrtbase
It's not allowed to link directly to the library (even though it exists).

For UCRT builds not packaged by MSVC vcruntime140* will not be found. It's not
a problem for UWP/UWP apps that are always built with MSVC.

It's a problem for UCRT builds that we might want to do in the future without
MSVC at all. We will need a replacement library for the calls we use. Some are
currently linked with a api-ms-win-crt-private DLL.