opengl: fragment_shaders: fix TexSize0 usage
In the case we're using GL_TEXTURE_RECTANGLE, we need to convert the
normalized texture coordinates into pixel coordinates using TexSize0.
This was done in the case of yuv (cf. is_yuv) in general but not when
is_yuv == false, which leads to failure when fetching uniform
locations on MacOSX.

In particular, this can be tested by forcing the chroma through the
--videotoolbox-cvpx-chroma=BGRA option on MacOSX.

Signed-off-by: Alexandre Janniaux <ajanni@videolabs.io>