contrib: gmp: don't use the cross compiler to test the build compiler
To build native tools it tries to use $CC when cross-compilation is detected,
which is most likely wrong.

This patch already exists in nettle that contains a version of gmp. But it's
not in the upstream version of gmp.

(cherry picked from commit 58e68e2f2ddaadc41c9ec4998df01b3daea7edde) (rebased)

rebased:
- the masyer branch has an asm fix detection patch

Signed-off-by: Steve Lhomme <robux4@ycbcr.xyz>