opengl: fix vertical alignment on control events
The video can be top-aligned (--align=4) or bottom-aligned (--align=8)
in the window.

In OpenGL, the vertical alignment is reversed, because the default
OpenGL convention orients the Y axis upwards while VLC convention
orients it downwards.

It was correctly handled for some control events (e.g.
CHANGE_DISPLAY_SIZE), but not for others (e.g. CHANGE_SOURCE_CROP).

Signed-off-by: Alexandre Janniaux <ajanni@videolabs.io>