mkv: ebml_dispatcher: fix ebmlid check
std::lower_bound doesn't return an iterator to the std::end value if the
element is not found, but like its name says, to the first value that is
greater or equal than the value wanted, meaning it could be the next one
if no processor has been bound to the ebml id sent to this dispatcher.

Instead of asserting, it should really be checking whether the ebmlid
matches.

Regression from c764461180d70d1c9fa81e72cd7ad9d9b289eea6.