image: forward video context when converting
When configuring the video converter filter chain, forward the video
context of the input picture into the chain. It allows starting the
correct conversion chain when the input picture has been made within a
video context (like VideoToolbox decoding) and need it to convert.

In particular, this commit fixes snapshot on MacOSX with VideoToolbox.