package/win32: build.sh: add a -o option to specify where to install vlc
This is convenient to specify the output rather than relying on the _win32
location. It avoids copying files from _win32 to whatever the needed location
is (in the libvlc UWP script for example).

This may also be useful to build in WSL2 on the Linux filesystem and install
on the Windows filesystem. It may be faster than building with WSL1 on the
Windows filesystem.