opengl: insert shader headers at compile time
The fields were used to move the dependency to OpenGL/GLES to runtime
and avoid opengl_tex_converter_t to be compiled twice on linux. Since
they were split into interop and samplers and that the interop doesn't
generate GLSL, the runtime dependency is not useful anymore.

It is not used for dynamic adaptation of shaders to the available GLSL
version too, so remove it for now.

Somehow revert of 524a003fb01e75f81f68ab61bd55b43e97c49fe9

Signed-off-by: Alexandre Janniaux <ajanni@videolabs.io>