opengl: properly handle interop fmt changes
The interop may modify its format for two different reasons:
 - to request a converter from the core
 - to describe its output format to the sampler (internal changes of
   software chroma, actual orientation, etc.)

The internal changes *must not* be reported to the core, to avoid an
unwanted insertion of a converter.

This was initially managed via a hack: all fields of the interop format
were copied back to the core format, except the orientation (because it
was the main "internal" change).

This hack was not sufficient due to internal chroma changes (RGB or YUV
fallbacks), which were wrongly reported to the core, leading to
unexpected insertion of converters. Commit
b79c2b224910d6987a3284ecaafeb4abb32a9a69 attempted to fix the problem by
simply not reporting any format changes to the core.

But interop_vdpau requires to change the input chroma (from
VLC_CODEC_VDPAU_VIDEO_42x to VLC_CODEC_VDPAU_OUTPUT). Since it was not
reported to the core anymore, the pictures received were not in the
expected format, causing a segfault.

To properly fix the issue, expose both an input format (reported to the
core) and an output format (for internal changes) on the interop.

This also removes the need for a "software format", which is now just a
specific property of the output format.

Signed-off-by: Francois Cartegnie <fcvlcdev@free.fr>